Add xbee daemon
The xbee daemon connects to an XBee (on /dev/ttyUSBx) and listens on one or
more local tcp sockets.

Any received packets from the XBee are decoded, re-encoded as JSON strings,
and sent to all connected clients. Any JSON strings received from a client
are decoded and sent to the XBee.