Introduce random errors in frame to test error handling
Introduce a small rate of byte errors in received data. The error
could occur in any byte or bytes of a frame. An erroneous frame
needs to be caught and the error reported to STDERR with enough
detail to reproduce what happened (show all the bytes received,
not find out what the correct packet should have been).